feat: Add user authentication functionality
This commit adds the necessary dependencies and
code changes to enable user authentication in the
YelpCamp project. The `passport`, `passport-local`,
and `passport-local-mongoose` packages are added to
the `package.json` file. The `User` model is created
in the `models/user.js` file to define the user schema
and methods for authentication. Middleware functions
for checking user authentication and storing the return
URL are added to the `middleware.js` file. The login
view is created in the `views/users/login.ejs` file to
allow users to log in to their accounts.