- finish Section56: Cluster Map - feat: Add map to campground index page
This commit adds a map to the campground index page
in the YelpCamp project. The map is displayed using
the Mapbox SDK and shows the locations of the campgrounds.
The map is rendered in a div with the id "map" and has a height of 500px.
The campgrounds are passed to the map as a JSON object and
are clustered using the clusterMap.js script. This enhancement
improves the user experience by providing a visual representation
of the campground locations.