- finish Section57 - feat: Update map display on campground index page
This commit updates the map display on the campground
index page in the YelpCamp project. The map is now rendered
in a div with the id "cluster-map" instead of "map". The map
style has also been changed to "mapbox://styles/mapbox/dark-v11".
These changes improve the visual representation of the campground
locations and enhance the user experience.